requirements.txt
selenium>=4.15.0
webdriver-manager>=4.0.1
google-genai
python-dotenv
Pillow>=10.0.0